+++ Sources/Views/MessagesCollectionView.swift
@@ -40,8 +40,8 @@ open class MessagesCollectionView: UICollectionView {
 
     private var indexPathForLastItem: IndexPath? {
 
-        let lastSection = numberOfSections > 0 ? numberOfSections - 1 : 0
-        guard lastSection > 0, numberOfItems(inSection: lastSection) > 0 else { return nil }
+        let lastSection = numberOfSections - 1
+        guard lastSection >= 0, numberOfItems(inSection: lastSection) > 0 else { return nil }
         return IndexPath(item: numberOfItems(inSection: lastSection) - 1, section: lastSection)
 
     }
@@ -65,7 +65,20 @@ open class MessagesCollectionView: UICollectionView {
 
     public func scrollToBottom(animated: Bool = false) {
         guard let indexPath = indexPathForLastItem else { return }
-        scrollToItem(at: indexPath, at: .bottom, animated: animated)
+        
+        let collectionViewContentHeight = collectionViewLayout.collectionViewContentSize.height
+        let isContentTooSmall = (collectionViewContentHeight < bounds.height * 2)
+        
+        if isContentTooSmall {
+            performBatchUpdates(nil) { _ in
+                self.scrollRectToVisible(CGRect(0.0, collectionViewContentHeight - 1.0, 1.0, 1.0), animated: animated)
+            }
+            return
+        }
+        
+        performBatchUpdates(nil) { _ in
+            self.scrollToItem(at: indexPath, at: .bottom, animated: animated)
+        }
     }
 
 }