+++ CHANGELOG.md
@@ -12,6 +12,14 @@ The changelog for `MessageKit`. Also see the [releases](https://github.com/Messa
 attributes to be set outside of the cell.
 [#397](https://github.com/MessageKit/MessageKit/pull/397) by [@SD10](https://github.com/sd10).
 
+### Fixed
+
+- Fixed `indexPathForLastItem` bug when `numberOfSections` equal to 1. 
+[#395](https://github.com/MessageKit/MessageKit/issues/395) by [@zhongwuzw](https://github.com/zhongwuzw).
+
+- Fixed `scrollToBottom(animated:)` not work in some situations.
+[#395](https://github.com/MessageKit/MessageKit/issues/395) by [@zhongwuzw](https://github.com/zhongwuzw).
+
 ### Changed
 
 - **Breaking Change** The `MessageLabel` properties `addressAttributes`, `dateAttributes`, `phoneNumberAttributes`,
@@ -35,6 +43,11 @@ and `messageLabel` for `TextMessageCell`.
 
 - **Breaking Change**  Adds a new view `contentView` of type `UIView` to the MessageInputBar to hold the main subviews of the `MessageInputBar`. Reduces complexity of constraints for easier testing/debugging.
 [#384](https://github.com/MessageKit/MessageKit/pull/384) by [@nathantannar4](https://github.com/nathantannar4).
+
+### Removed
+
+- **Breaking Change** Removed `scrollsToBottomOnFirstLayout` flag of `MessagesViewController`.
+[#395](https://github.com/MessageKit/MessageKit/pull/395) by [@zhongwuzw](https://github.com/zhongwuzw).
 
 ## [[Prerelease] 0.11.0](https://github.com/MessageKit/MessageKit/releases/tag/0.11.0)